// src/types/mainEnums.ts
export enum marital_status {
  single = "single",
  married = "married",
  divorced = "divorced",
  widowed = "widowed",
}
export enum gender {
  male = "male",
  female = "female",
  other = "other",
}
export enum days {
  monday = "Monday",
  tuesday = "Tuesday",
  wednesday = "Wednesday",
  thursday = "Thursday",
  friday = "Friday",
  saturday = "Saturday",
  sunday = "Sunday",
}
export enum memberType {
  remote = "Remote",
  office = "Office",
}
export enum status {
  permament = "Permament",
}
export enum jobTypes {
  active = "Active",
  inactive = "Inactive",
  completed = "Completed",
}
export enum processStatus {
  completed = "Completed",
  inProcess = "In Process",
}
export enum attendanceStatus {
  onTime = "On Time",
  late = "Late",
  absent = "Absent",
}
export enum candidateStatus {
  selected = "Selected",
  inProcess = "In Process",
  rejected = "Rejected",
}
export enum payrollStatus {
  completed = "Completed",
  pending = "Pending",
}
export enum leaveStatus {
  approved = "Approved",
  pending = "Pending",
  reject = "Reject",
}